requirement.txt
altair~=4.2.0
dnspython~=2.2.0
Flask~=2.0.3
gunicorn~=20.1.0
joblib~=1.1.0
pandas~=1.4.1
pymongo~=4.0.1
python-dotenv~=0.19.2
pytz~=2021.3
scikit-learn~=1.0.2